Break out ShardedBAMWriting.java into multiple files.
